--- src/import-cache/policy.go
@@ -1,9 +1,0 @@
-package cache
-
-type CacheMatcher interface {
-	Match(servers []*ForkServer, pkgs []string) (*ForkServer, []string, bool)
-}
-
-type CacheEvictor interface {
-	Evict(servers []*ForkServer) error
-}
